+++ x/core/02_post_dispatch/client/cli/tx.go
@@ -6,8 +6,8 @@ import (
 )
 
 var (
+	newOwner          string
 	renounceOwnership bool
-	yes               bool
 )
 
 func GetTxCmd() *cobra.Command {