+++ x/core/client/cli/tx_mailbox.go
@@ -122,6 +122,11 @@ func CmdSetMailbox() *cobra.Command {
 		Short: "Update a Hyperlane Mailbox",
 		Args:  cobra.ExactArgs(1),
 		PreRunE: func(cmd *cobra.Command, args []string) error {
+			yes, err := cmd.Flags().GetBool("yes")
+			if err != nil {
+				return err
+			}
+
 			if renounceOwnership && !yes {
 				fmt.Print("Are you sure you want to renounce ownership? This action is irreversible. (yes/no): ")
 				var response string
@@ -187,7 +192,6 @@ func CmdSetMailbox() *cobra.Command {
 	cmd.Flags().StringVar(&requiredHook, "required-hook", "", "set updated requiredHook")
 	cmd.Flags().StringVar(&newOwner, "new-owner", "", "set updated owner")
 	cmd.Flags().BoolVar(&renounceOwnership, "renounce-ownership", false, "renounce ownership")
-	cmd.Flags().BoolVarP(&yes, "yes", "y", false, "auto-confirm all prompts")
 
 	flags.AddTxFlagsToCmd(cmd)