+++ x/warp/client/cli/tx.go
@@ -13,7 +13,6 @@ var (
 	maxFee             string
 	newOwner           string
 	renounceOwnership  bool
-	yes                bool
 )
 
 func GetTxCmd() *cobra.Command {