+++ CHANGELOG.md
@@ -16,9 +16,9 @@ An '!' indicates a state machine breaking change.
 
 ## Unreleased
 
-### Improvements
+### Features
 
-- [#98](https://github.com/bcp-innovations/hyperlane-cosmos/pull/98) Add renounce ownership functionalities and validate new owner
+- ! [#98](https://github.com/bcp-innovations/hyperlane-cosmos/pull/98) Add renounce ownership functionalities and validate new owner
 
 ## [v1.0.0-beta0](https://github.com/bcp-innovations/hyperlane-cosmos/releases/tag/v1.0.0-beta0) - 2025-03-27